third_party/blink/web_tests/external/wpt/svg/types/scripted/SVGGraphicsElement.getBBox-09-expected.txt

This is a testharness.js-based test.
[FAIL] stroked zero-width rect: not rendered, fill bbox is all zeros
  assert_approx_equals: x expected 0 +/- 0.01 but got 10
[FAIL] stroked zero-height rect: not rendered, fill bbox is all zeros
  assert_approx_equals: x expected 0 +/- 0.01 but got 10
[FAIL] stroked zero-dim rect: not rendered, fill bbox is all zeros
  assert_approx_equals: x expected 0 +/- 0.01 but got 10
[FAIL] stroked zero-radius circle: not rendered, fill bbox is all zeros
  assert_approx_equals: x expected 0 +/- 0.01 but got 10
[FAIL] stroked horizontal line: stroke-width wide
  assert_approx_equals: y expected 18 +/- 0.01 but got 20
[FAIL] stroked vertical line: stroke-width wide
  assert_approx_equals: x expected 8 +/- 0.01 but got 10
[FAIL] stroked zero-width rect: not rendered, stroke bbox is all zeros
  assert_approx_equals: x expected 0 +/- 0.01 but got 10
[FAIL] stroked zero-height rect: not rendered, stroke bbox is all zeros
  assert_approx_equals: x expected 0 +/- 0.01 but got 10
[FAIL] stroked zero-dim rect: not rendered, stroke bbox is all zeros
  assert_approx_equals: x expected 0 +/- 0.01 but got 10
[FAIL] stroked zero-radius circle: not rendered, stroke bbox is all zeros
  assert_approx_equals: x expected 0 +/- 0.01 but got 10
[FAIL] stroked zero-length path segment with round linecap: stroke-width wide
  assert_approx_equals: x expected 8 +/- 0.01 but got 10
[FAIL] stroked ellipse with zero rx: not rendered, stroke bbox is all zeros
  assert_approx_equals: x expected 0 +/- 0.01 but got 10
[FAIL] stroked horizontal path with butt linecap: stroke-width wide
  assert_approx_equals: y expected 18 +/- 0.01 but got 20
[FAIL] zero-length line with fill:false stroke:false returns empty bbox
  assert_approx_equals: x expected 0 +/- 0.01 but got 10
Harness: the test ran to completion.
